requirement.txt
langgraph
langchain-core
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
streamlit
python-dotenv


